# Removed auto-decoding from the bool version of std.conv.parse