@JackStouffer 2017-02-12
Auto-merge toggled on